refactor: my position card tooltips
