--- packages/web/src/models/position/info/position-balance-info.ts
@@ -1,8 +1,0 @@
-import { TokenModel } from "@models/token/token-model";
-
-export interface PositionBalanceInfo {
-  token: TokenModel;
-  balance: number;
-  balanceUSD: number;
-  percent: string;
-}
