--- packages/web/src/models/position/info/position-claim-info.ts
@@ -1,11 +1,0 @@
-import { RewardType } from "@constants/option.constant";
-import { TokenModel } from "@models/token/token-model";
-
-export interface PositionClaimInfo {
-  rewardType: RewardType;
-  token: TokenModel;
-  claimableAmount: number | null;
-  claimableUSD: number | null;
-  accumulatedRewardOf1d: number | null;
-  accumulatedRewardOf1dUsd: number | null;
-}
