--- packages/web/src/models/position/info/position-reward-info.ts
@@ -1,11 +1,0 @@
-import { TokenModel } from "@models/token/token-model";
-
-export interface PositionRewardInfo {
-  claimableAmount: number;
-  token: TokenModel;
-  balance: number;
-  balanceUSD: number;
-  claimableUSD: number | null;
-  accumulatedRewardOf1d: number | null;
-  accumulatedRewardOf1dUsd: number | null;
-}
