+++ packages/web/src/views/pool/pool-detail/components/my-liquidity/my-detailed-position-card/MyDetailedPositionCard.styles.ts
@@ -369,49 +369,6 @@ export const TooltipContent = styled.div`
   }
 `;
 
-export const RewardsContent = styled.div`
-  ${mixins.flexbox("column", "flex-start", "space-between")};
-  gap: 8px;
-  width: 268px;
-  ${fonts.body12};
-  ${media.mobile} {
-    gap: 4px;
-    ${fonts.p2};
-  }
-  .list {
-    ${mixins.flexbox("row", "center", "space-between")};
-    width: 100%;
-    padding: 4px 0px;
-    .coin-info {
-      ${mixins.flexbox("row", "center", "flex-start")};
-      gap: 8px;
-      flex-shrink: 0;
-      .token-logo {
-        width: 20px;
-        height: 20px;
-      }
-    }
-  }
-  .title {
-    color: ${({ theme }) => theme.color.text04};
-  }
-  .note {
-    color: ${({ theme }) => theme.color.text04};
-    ${fonts.p4}
-  }
-  .content {
-    color: ${({ theme }) => theme.color.text02};
-  }
-  p {
-    ${fonts.p4};
-    color: ${({ theme }) => theme.color.text04};
-  }
-  .divider {
-    width: 100%;
-    border-top: 1px solid ${({ theme }) => theme.color.border01};
-  }
-`;
-
 export const TooltipDivider = styled.div`
   ${mixins.flexbox("column", "center", "flex-start")};
   height: 1px;
